python_scripts/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
albumentations==1.4.3
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
blosc2==3.11.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chex==0.1.90
chumpy==0.70
click==8.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.0
dadapy==0.3.3
debugpy==1.8.16
decorator==5.2.1
deeplabcut==3.0.0rc13
defusedxml==0.7.1
dlclibrary==0.0.11
einops==0.8.1
etils==1.13.0
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.19.1
filterpy==1.4.5
flatbuffers==25.2.10
flax==0.8.5
fonttools==4.59.2
fqdn==1.5.1
fsspec==2025.7.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.0rc2
humanize==4.14.0
idna==3.10
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
imgaug==0.4.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jax==0.4.30
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json-tricks==3.17.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.11.3
kiwisolver==1.4.9
lark==1.3.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.45.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.3
mmpose==1.3.2
mpi4py==4.1.0
mpmath==1.3.0
msgpack==1.1.2
munkres==1.1.4
namex==0.1.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.5
notebook_shim==0.2.4
numba==0.62.0
numexpr==2.14.1
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optax==0.2.5
optree==0.17.0
orbax-checkpoint==0.6.4
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
playsound==1.3.0
pluggy==1.6.0
polars==1.35.1
polars-runtime-32==1.35.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.10
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8
stack-data==0.6.3
statsmodels==0.14.5
sympy==1.14.0
tables==3.10.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorstore==0.1.79
termcolor==3.1.0
terminado==0.18.1
tf_keras==2.20.1
threadpoolctl==3.6.0
tifffile==2025.10.16
timm==1.0.22
tinycss2==1.4.0
tokenizers==0.22.1
toolz==1.1.0
torch==2.8.0
torchao==0.13.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers@e11a00a16f925b7d3b52f5007bdce3464edb361f
typer-slim==0.19.2
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.223
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
xtcocotools==1.14.3
yt-dlp==2025.8.27
zipp==3.23.0
